LayoutTests/accessibility/placeholder-expected.txt


This test makes sure that the AXTitle will fall back to placeholder text if nothing else is available

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fieldElement.title is 'AXTitle: search'
PASS successfullyParsed is true

TEST COMPLETE
